поправила пути модулей, поправила название js файла
