--- js/modal_big_pictures.js
@@ -1,1 +1,0 @@
-
